move into mixin class
